+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/demand/logic/services/DemandService.java
@@ -19,10 +19,6 @@ SPDX-License-Identifier: Apache-2.0
 */
 package org.eclipse.tractusx.puris.backend.demand.logic.services;
 
-import java.time.Instant;
-import java.time.LocalDate;
-import java.time.ZoneOffset;
-import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;