+++ CHANGELOG.md
@@ -11,10 +11,14 @@ The following Changelog lists the changes. Please refer to the [documentation](d
 
 The **need for configuration updates** is **marked bold**.
 
+### Added
+
+* Implement Excel imports through API ([#897](https://github.com/eclipse-tractusx/puris/pull/897/))
+
 ### Changed
 
 * Use the frontend's existing keycloak authentication to authorize backend calls ([#896](https://github.com/eclipse-tractusx/puris/pull/896)) (**moved IDP configuration and IDP no longer optional**)
-* Implement Excel imports through API ([#897](https://github.com/eclipse-tractusx/puris/pull/897/))
+* Rework validation logic to return detailed error messages ([#898](https://github.com/eclipse-tractusx/puris/pull/898))
 
 ## [v3.1.0](https://github.com/eclipse-tractusx/puris/releases/tag/3.1.0)